# internal/apidoc/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    name: Aegis
    url: https://github.com/wisonwang/Aegis
  description: Aegis — AI Data Supply Gateway. 受治理的 DataAPI + 面向 AI Agent 的 MCP 服务。默认拒绝：每条查询都经过权限引擎重写与脱敏。
  license:
    name: MIT
  termsOfService: https://github.com/wisonwang/Aegis
  title: Aegis DataAPI
  version: "1.0"
paths:
...
  /admin/api/datasources/{id}/masks:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Masks
      tags:
      - masks
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Upsert Mask
      tags:
      - masks
  /admin/api/datasources/{id}/masks/{mask}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: mask
        in: path
        name: mask
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Mask
      tags:
      - masks
  /admin/api/datasources/{id}/masks/recommend:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Recommend Masks
      tags:
      - masks
...
  /admin/api/datasources/{id}/semantics:
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin List Semantics
      tags:
      - semantics
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Upsert Semantic
      tags:
      - semantics
  /admin/api/datasources/{id}/semantics/{sem}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: sem
        in: path
        name: sem
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: admin Delete Semantic
      tags:
      - semantics
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <JWT> — 通过 POST /api/v1/login 获取"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"